# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AccountResponse {
  id: string;
  email: string;
  fullName: string;
  createdAt: string;
  updatedAt: string;
}

export type ListAccountsResponse = AccountResponse[];

export interface CreateAccountBody {
  email: string;
  /** @minLength 1 */
  fullName: string;
}

export interface UpdateAccountBody {
  email?: string;
  /** @minLength 1 */
  fullName?: string;
}

export interface PortfolioResponse {
  id: string;
  accountId: string;
  name: string;
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ListPortfoliosResponse = PortfolioResponse[];

export interface CreatePortfolioBody {
  accountId: string;
  /** @minLength 1 */
  name: string;
  description?: string | null;
}

export interface UpdatePortfolioBody {
  /** @minLength 1 */
  name?: string;
  description?: string | null;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  stock: 'stock',
  etf: 'etf',
  crypto: 'crypto',
  bond: 'bond',
  commodity: 'commodity',
  other: 'other',
} as const;

export interface HoldingResponse {
  id: string;
  portfolioId: string;
  ticker: string;
  assetType: AssetType;
  quantity: string;
  avgCostBasis?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ListHoldingsResponse = HoldingResponse[];

export interface CreateHoldingBody {
  portfolioId: string;
  /** @minLength 1 */
  ticker: string;
  assetType: AssetType;
  quantity: string;
  avgCostBasis?: string | null;
}

export interface UpdateHoldingBody {
  quantity?: string;
  avgCostBasis?: string | null;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  buy: 'buy',
  sell: 'sell',
  dividend: 'dividend',
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export interface TransactionResponse {
  id: string;
  portfolioId: string;
  ticker: string;
  transactionType: TransactionType;
  quantity?: string | null;
  pricePerUnit: string;
  totalAmount: string;
  notes?: string | null;
  transactedAt: string;
  createdAt: string;
}

export type ListTransactionsResponse = TransactionResponse[];

export interface CreateTransactionBody {
  portfolioId: string;
  /** @minLength 1 */
  ticker: string;
  transactionType: TransactionType;
  quantity?: string | null;
  pricePerUnit: string;
  totalAmount: string;
  notes?: string | null;
  transactedAt: string;
}

export type ListPortfoliosParams = {
/**
 * Filter by account ID
 */
accountId?: string;
};

export type ListHoldingsParams = {
/**
 * Filter by portfolio ID
 */
portfolioId?: string;
};

export type ListTransactionsParams = {
/**
 * Filter by portfolio ID
 */
portfolioId?: string;
/**
 * Filter by ticker symbol
 */
ticker?: string;
};