# lib/api-zod/src/generated/types/assetType.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  stock: 'stock',
  etf: 'etf',
  crypto: 'crypto',
  bond: 'bond',
  commodity: 'commodity',
  other: 'other',
} as const;