# lib/api-zod/src/generated/types/createHoldingBody.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetType } from './assetType';

export interface CreateHoldingBody {
  portfolioId: string;
  /** @minLength 1 */
  ticker: string;
  assetType: AssetType;
  quantity: string;
  avgCostBasis?: string | null;
}