# lib/api-zod/src/generated/types/listAccountsResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountResponse } from './accountResponse';

export type ListAccountsResponse = AccountResponse[];